# openapi/oh.yaml
openapi: 3.0.1
info:
  title: Open Hospital API Documentation
  description: Open Hospital API Documentation
  contact:
    name: ApiInfo.DEFAULT_CONTACT
  license:
    name: GPL-3.0 license
    url: https://github.com/informatici/openhospital-api?tab=GPL-3.0-1-ov-file#readme
  version: 0.1.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: [ ]
paths:
...
  /usergroups/{group_code}:
    get:
      tags:
        - User Groups
      operationId: getUserGroup
      parameters:
        - name: group_code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGroupDTO"
      security:
        - bearerAuth: [ ]
    put:
      tags:
        - User Groups
      operationId: updateUserGroup
      parameters:
        - name: group_code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserGroupDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGroupDTO"
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - User Groups
      operationId: deleteGroup
      parameters:
        - name: group_code
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
      security:
        - bearerAuth: [ ]
  /usergroups/{group_code}/permissions:
    put:
      tags:
        - User Groups
      operationId: updateGroupPermissions
      parameters:
        - name: group_code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupPermissionsDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PermissionDTO"
      security:
        - bearerAuth: [ ]
    patch:
      tags:
        - User Groups
      operationId: replaceGroupPermissions
      parameters:
        - name: group_code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupPermissionsDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PermissionDTO"
      security:
        - bearerAuth: [ ]
...
  /usergroups:
    get:
      tags:
        - User Groups
      operationId: getUserGroups
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserGroupDTO"
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - User Groups
      operationId: newUserGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserGroupDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGroupDTO"
      security:
        - bearerAuth: [ ]
  /usergroups/{group_code}/permissions/{id}:
    post:
      tags:
        - User Groups
      operationId: assignPermission
      parameters:
        - name: group_code
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: integer
                format: int32
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - User Groups
      operationId: revokePermission
      parameters:
        - name: group_code
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "204":
          description: No Content
      security:
        - bearerAuth: [ ]
...
  /permissions/{id}:
    get:
      tags:
        - Permissions
      operationId: retrievePermissionById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionDTO"
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
...
    PermissionDTO:
      required:
        - description
        - id
      type: object
      properties:
        name:
          type: string
        id:
          type: integer
          description: The permission id
          format: int32
        description:
          maxLength: 255
          type: string
          description: The permission description
      description: List of group's permissions
...
    UserGroupDTO:
      required:
        - code
      type: object
      properties:
        code:
          maxLength: 50
          type: string
          description: Name of the group (must be unique)
          example: labo
        desc:
          maxLength: 128
          type: string
          description: The description of the group
          example: Staff members working in the laboratory
        deleted:
          type: boolean
          description: Whether the group has been soft deleted or not
          example: false
        permissions:
          type: array
          description: List of group's permissions
          items:
            $ref: "#/components/schemas/PermissionDTO"
      description: The user's group
...
    GroupPermissionsDTO:
      type: object
      properties:
        permissions:
          type: array
          description: List of permissions' ids
          example:
            - 48
            - 24
          items:
            type: integer
            description: List of permissions' ids
            format: int32
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT